+++ src/main/typescript/lang.d.ts
@@ -369,6 +369,26 @@ export interface LangSchema {
       effect: QuantityNames;
       /** Labels related to generic Misc Items */
       item: QuantityNames;
+      /** Labels used in the initial character setup app */
+      initialCharacterSetup: {
+        /** The label for the open button on the character sheet */
+        openButton: string;
+        /**
+         * The title for the window
+         *
+         * Parameters:
+         * - name: the name of the character
+         * @pattern (?=.*\{name\})
+         */
+        title: string;
+        /** Labels relating to the initial character setup */
+        messages: {
+          /** A message for when the user spent too few SPECIAL points */
+          tooFewSpecialPointsSpent: string;
+          /** A message for when the user spent too many SPECIAL points */
+          tooManySpecialPointsSpent: string;
+        };
+      };
       /** Different system messages. */
       messages: {
         /**